split docker.go
